Move paralysis speed test to the pokemon tests
