+++ tests/engine_tests/pokemon_engine_test.py
@@ -90,37 +90,8 @@ def test_heal():
         p_eng.game_state["player1"]["active"].max_hp
 
 
-def test_status():
-    """Test that the engine does status moves properly"""
-    test_speed_paralyze()
-
-
-def test_speed_paralyze():
-    """
-    Test paralysis speed drop.
-    
-    Spinda should outspeed exploud and kill it.
-    """
-    spinda = Pokemon(name="spinda", moves=["tackle"])
-    exploud = Pokemon(name="exploud", moves=["tackle"])
-    spinda.current_hp = 1
-    exploud.current_hp = 1
-    exploud.status = "par"
-
-    player1 = PokemonAgent([spinda])
-    player2 = PokemonAgent([exploud])
-
-    p_eng = PokemonEngine()
-    p_eng.run(player1, player2)
-    outcome = p_eng.win_condition_met()
-
-    assert outcome["finished"]
-    assert outcome["winner"] == 1
-
-
 test_run()
 test_run_multiple_moves()
 test_run_multiple_pokemon()
 test_run_infinite()
 test_heal()
-test_status()
